requirements.txt
streamlit==1.29.0
Pillow==10.1.0
streamlit-lottie==0.0.5
requests==2.31.0
streamlit-extras==0.3.6
streamlit-particles==0.0.1
